refactor(frontend): share the week day cell instead of copying it
Sonar was right to fail this: 14.5% new duplicated lines against a 3%
gate. I converged the Tasks date strip by copying the Appointments
treatment verbatim, which is the same mistake that produced the
divergence in the first place - two copies of one design, free to drift.

The column is CalendarWeekDayCell now and both week headers render it,
so the next change lands on both calendars at once. Rendering is
unchanged: the story screenshots are identical before and after, and the
contrast matrix is still clean.

916 suites / 11275 tests.
